% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{ev}
\alias{ev}
\title{Calculate Expected Values of a Model}
\usage{
ev(param, ...)
}
\arguments{
\item{param}{Paramters (coefficients) of the model. Often in the same format as returned by \code{\link{coef}}.}

\item{...}{extra arguments}
}
\description{
This method calculates the expected value of common models
given known parameter values. Although it is generally easy to calculate expected values
of a model after fitting it using the function \code{predict},
there is often not an easy way to do this when the parameters
of the model are known, rather than estimated.
}

