% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{fdfx}
\alias{afdfx}
\alias{fdfx}
\title{(Average) Partial Effects}
\usage{
fdfx(x, data1, data2, delta, ...)

afdfx(x, data1, data2, delta, ...)
}
\arguments{
\item{x}{A model object}

\item{data1, data2}{Model data}

\item{delta}{Width of difference. See Details.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The partial effects are the effects of discrete change
in a variable.
}

