% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-glm.R
\name{glm-methods}
\alias{afdfx.glm}
\alias{fdfx.glm}
\alias{glm-methods}
\alias{simev.glm}
\alias{simfdfx.glm}
\alias{simpar.glm}
\title{Methods for \code{glm} Models}
\usage{
\method{fdfx}{glm}(x, data1, data2, delta = 1, n = 1000L, confint = 0.95,
  response = TRUE, ...)

\method{afdfx}{glm}(x, data1, data2, delta = 1, n = 1000L, confint = 0.95,
  response = TRUE, weights = NULL, ...)

\method{simev}{glm}(x, data = stats::model.frame(x), response = TRUE,
  n = 1000L, ...)

\method{simfdfx}{glm}(x, data1, data2, n = 1L, delta = 1,
  response = FALSE, ...)

\method{simpar}{glm}(x, n = 1L, V = NULL, ...)
}
\arguments{
\item{x}{object}

\item{data1, data2, data}{Data}

\item{delta}{Size of the difference}

\item{n}{Number of iterations}

\item{confint}{Confidence interval level}

\item{response}{If response is true, then partial effects are on the
the scale of the response variable. If false, then the partial effects
are on the scale of the linear predictors.}

\item{...}{further arguments passed to or from other methods.}

\item{weights}{Weights to apply to the data}

\item{V}{The variance-covariance matrix of the coefficients. This arguments
allows for the substitution of "robust" covariance matrices.}
}
\description{
Methods for \code{glm} Models
}

