% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-lm.R
\name{lm-methods}
\alias{avg_partialfx.lm}
\alias{lm-methods}
\alias{partialfx.lm}
\alias{postsim.lm}
\alias{postsim_partialfx.lm}
\alias{postsimev.lm}
\alias{postsimy.lm}
\title{Methods for \code{lm} Models}
\usage{
\method{postsim}{lm}(x, n = 1L, ...)

\method{postsim_partialfx}{lm}(x, data1, data2, n = 1L, delta = 1, ...)

\method{partialfx}{lm}(x, data1, data2, delta = 1, n = 1000L,
  confint = 0.95, ...)

\method{avg_partialfx}{lm}(x, data1, data2, delta = 1, n = 1000L,
  confint = 0.95, weights = NULL, ...)

\method{postsimev}{lm}(x, data = stats::model.frame(x), n = 1000L, ...)

\method{postsimy}{lm}(x, n = 1L, data = stats::model.frame(x), ...)
}
\arguments{
\item{x}{object}

\item{n}{Number of iterations}

\item{...}{further arguments passed to or from other methods.}

\item{data1, data2, data}{Data}

\item{delta}{Size of the difference}

\item{confint}{Confidence interval level}

\item{weights}{Weights to apply to the data}
}
\description{
Methods for \code{lm} Models
}

