% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{marfx}
\alias{avg_marfx}
\alias{marfx}
\title{Calculate (Average) Marginal Effects of a Model}
\usage{
marfx(x, ...)

avg_marfx(x, ...)
}
\arguments{
\item{x}{A model object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
The partial effects are the effects of discrete change
in a variable.
}

