% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{numdiff_stepsize}
\alias{numdiff_stepsize}
\title{Numerical Differentiation Step-size}
\usage{
numdiff_stepsize(x)
}
\arguments{
\item{x}{The value of \eqn{x} at which the derivative is being evaluated.}
}
\description{
Step-size for one-sided numerical differentiation to
\deqn{h = \max(|x|, 1) \sqrt{\epsilon}}{h = max(|x|, 1)sqrt(epsilon)}
where \eqn{\epsilon}{epsilon} is machien precision.
}
\references{
Miranda, Mario J. and Paul L. Fackler. 2002. Applied Computational
  Economics and Finance. p. 103.
}

