% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{ordered}
\alias{median.ordered}
\alias{ordered}
\alias{quantile.ordered}
\title{Methods for ordered factors}
\usage{
\method{median}{ordered}(x, na.rm = FALSE)

\method{quantile}{ordered}(x, probs = seq(0, 1, 0.25), na.rm = FALSE,
  left = TRUE, ...)
}
\description{
Methods for ordered factors
}

