% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{postsim_partialfx}
\alias{postsim_partialfx}
\title{Posterior Simulation of Partial Effects}
\usage{
postsim_partialfx(x, data1, data2, n, delta, ...)
}
\arguments{
\item{x}{Model object}

\item{data1, data2}{Data frames}

\item{n}{Number of iterations}

\item{delta}{Width of difference. See Details.}

\item{...}{Optional parameters}
}
\description{
Calculate the partial (discrete marginal) effects of a model.
}
\details{
This function calculates:
\deqn{(E(y|X_1)) - E(y|X_2)) / \delta, }{((E(y|X1) - E(y|X2)) / delta, }
where \eqn{\delta}{delta} is a difference to scale by.
}

