% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{postsimev}
\alias{postsimev}
\title{Simulate Expected Values from a Model}
\usage{
postsimev(x, ...)
}
\arguments{
\item{x}{A model object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Simulate Expected Values from a Model
}

