% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{postsimy}
\alias{postsimy}
\title{Simulate Responses from a Model}
\usage{
postsimy(x, ...)
}
\arguments{
\item{x}{A model object}

\item{...}{extra arguments}
}
\description{
Simulate Responses from a Model
}

