% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{simy}
\alias{simy}
\title{Simulate Responses from a Model}
\usage{
simy(x, ...)
}
\arguments{
\item{x}{A model object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Simulate Responses from a Model
}

