% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{partialfx}
\alias{avg_partialfx}
\alias{partialfx}
\title{Calculate (Average) Partial Effects of a Model}
\usage{
partialfx(x, ...)

avg_partialfx(x, ...)
}
\arguments{
\item{x}{A model object}

\item{...}{extra arguments}
}
\description{
The partial effects are the effects of discrete change
in a variable.
}

