% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{partialfx}
\alias{avg_partialfx}
\alias{partialfx}
\title{(Average) Partial Effects}
\usage{
partialfx(x, data1, data2, delta, ...)

avg_partialfx(x, data1, data2, delta, ...)
}
\arguments{
\item{x}{A model object}

\item{data1, data2}{Model data}

\item{delta}{Width of difference. See Details.}

\item{...}{Extra arguments}
}
\description{
The partial effects are the effects of discrete change
in a variable.
}

