% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfx.R
\name{postsim}
\alias{postsim}
\title{Simulate Parameters from a Model}
\usage{
postsim(x, ...)
}
\arguments{
\item{x}{A model object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Simulate Parameters from a Model
}

